% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linear_aft}
\alias{linear_aft}
\title{Generate data from linear log-normal AFT model.}
\usage{
linear_aft(n, p, sd_err, support_beta, censoring_rate)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of covariates.}

\item{sd_err}{Standard deviation of the error term, \eqn{\epsilon}.}

\item{support_beta}{Support of \eqn{\beta} (non-zero coefficients).}

\item{censoring_rate}{Censoring rate.}
}
\value{
linearAFT returns an object of class "survsim".  An object of class
       "survsim" is a list containing the following components.
   \itemize{
   \item X: matrix of covariates
   \item y: unobserved log-survival times
   \item t: unobserved survival times
   \item c: unobserved censoring times.
   \item obs_times: observed time on study (min(t,c)).
   \item delta: vector indicating whether observation is
   censored (1=failure, 0=censored).
   }
}
\description{
Generate data from linear log-normal AFT model.
}
\examples{
linear_aft(1000, 5, .1, 1, .3)
}

