% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linear_normal_aft}
\alias{linear_normal_aft}
\title{Generate data from linear log-normal AFT model.}
\usage{
linear_normal_aft(n, p, sd_err, support_beta, censoring_rate)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of covariates.}

\item{sd_err}{Standard deviation of the error term, \eqn{\epsilon}.}

\item{support_beta}{Support of \eqn{\beta} (non-zero coefficients).}

\item{censoring_rate}{Censoring rate.}
}
\value{
linear_aft returns an object of class "survsim".  An object of class
       "survsim" is a list containing the following components.
   \itemize{
   \item X: matrix of covariates
   \item y: unobserved log-survival times
   \item t: unobserved survival times
   \item c: unobserved censoring times.
   \item obs_times: observed time on study (min(t,c)).
   \item delta: vector indicating whether observation is
   censored (1=failure, 0=censored).
   }
}
\description{
Generate data from linear log-normal AFT model.
}
\details{
This function generates a data from a user specified
log normal AFT model.  The user specifies the sample size,
the number of covariates, support of beta, and censoring rate.
The elements of \eqn{X} are iid standard normal.
Let \eqn{\beta= (\beta_{supp},0)}.  The log survival times, \eqn{y} are
generated as follows \eqn{y = X\beta + \epsilon} where the elements of
\eqn{\epsilon} are iid normal with mean 0 and standard deviation
\code{sd_err}.
}
\examples{
linear_normal_aft(1000, 5, .1, 1, .3)
}

