% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linearAFT}
\alias{linearAFT}
\title{Generate data from linear log-normal AFT model.}
\usage{
linearAFT(n, p, sd_err, nonzero_beta, censoring_rate)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of covariates.}

\item{sd_err}{Standard error of \eqn{\epsilon}.}

\item{nonzero_beta}{Vector of non-zero coefficients.}

\item{censoring_rate}{Censoring rate.}
}
\value{
linearAFT returns an object of class "survsim".
}
\description{
Generate data from linear log-normal AFT model.
}
\examples{
linearAFT(1000, 5, .1, 1, .3)
}

