% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nonlinear_normal_aft}
\alias{nonlinear_normal_aft}
\title{Generate data from a general nonlinear log-normal AFT model.}
\usage{
nonlinear_normal_aft(n, p, sd_err, f, censoring_rate)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of covariates.}

\item{sd_err}{Standard deviation of the error term, \eqn{\epsilon}.}

\item{f}{True regression function for log survival times. The function
f should take a vector as its argument.}

\item{censoring_rate}{Censoring rate.}
}
\value{
linear_aft returns an object of class "survsim".  An object of class
       "survsim" is a list containing the following components.
   \itemize{
   \item X: matrix of covariates
   \item y: unobserved log-survival times
   \item t: unobserved survival times
   \item c: unobserved censoring times.
   \item obs_times: observed time on study \eqn{min(t,c)}.
   \item delta: vector indicating whether observation is
   censored (1=failure, 0=censored).
   }
}
\description{
Generate data from a general nonlinear log-normal AFT model.
}
\details{
This function generates a data from a user specified
log normal AFT model.  The user specifies the sample size,
the number of covariates, support of beta, and censoring rate.
The elements of \eqn{X} are iid standard normal.
Log survival times, \eqn{y_{i}} are
generated as follows \eqn{y_{i} = f(x_{i}) + \epsilon_{i}} where the
\eqn{\epsilon_{i}} are iid normal with mean 0 and standard deviation
\code{sd_err}.
}
\examples{
#Simple AFT model with an interaction.
f <- function(x){
 x[1] + x[2] + x[1]*x[2]
}
nonlinear_normal_aft(1000, 5, .1, f, .25)
}

